Move activity under eks
