+++ cmd/worker/eks.go
@@ -183,7 +183,7 @@ func registerEKSWorkflows(
 	// Node pool upgrade
 	eksworkflow2.NewUpdateNodePoolWorkflow(awsSessionFactory, cloudFormationFactory, processlog.New()).Register(worker)
 
-	awsworkflow.NewCalculateNodePoolVersionActivity().Register(worker)
+	eksworkflow.NewCalculateNodePoolVersionActivity().Register(worker)
 	eksworkflow2.NewUpdateNodeGroupActivity(awsSessionFactory, nodePoolTemplate, defaultNodeVolumeEncryption).Register(worker)
 	eksworkflow2.NewWaitCloudFormationStackUpdateActivity(awsSessionFactory).Register(worker)
