+++ internal/cluster/distribution/eks/eksprovider/workflow/activity_calculate_node_pool_version_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package awsworkflow
+package workflow
 
 import (
 	"fmt"
