+++ internal/cluster/distribution/eks/eksprovider/workflow/create_node_pool_workflow.go
@@ -25,7 +25,6 @@ import (
 	"github.com/banzaicloud/pipeline/internal/cluster"
 	"github.com/banzaicloud/pipeline/internal/cluster/distribution/eks"
 	"github.com/banzaicloud/pipeline/internal/cluster/distribution/eks/eksmodel"
-	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 	pkgcadence "github.com/banzaicloud/pipeline/pkg/cadence"
 	"github.com/banzaicloud/pipeline/pkg/cadence/worker"
 	sdkamazon "github.com/banzaicloud/pipeline/pkg/sdk/providers/amazon"
@@ -234,7 +233,7 @@ func (w CreateNodePoolWorkflow) Execute(ctx workflow.Context, input CreateNodePo
 		return err
 	}
 
-	nodePoolVersion, err := awsworkflow.CalculateNodePoolVersion(
+	nodePoolVersion, err := CalculateNodePoolVersion(
 		ctx, input.NodePool.Image, input.NodePool.VolumeEncryption, input.NodePool.VolumeSize, input.NodePool.SecurityGroups)
 	if err != nil {
 		return err
