+++ internal/cluster/distribution/eks/eksprovider/workflow/create_node_pool_workflow_test.go
@@ -33,7 +33,6 @@ import (
 	"github.com/banzaicloud/pipeline/internal/cluster/clusterworkflow"
 	"github.com/banzaicloud/pipeline/internal/cluster/distribution/eks"
 	"github.com/banzaicloud/pipeline/internal/cluster/distribution/eks/eksmodel"
-	"github.com/banzaicloud/pipeline/internal/cluster/infrastructure/aws/awsworkflow"
 )
 
 func TestCreateNodePool(t *testing.T) {
@@ -751,17 +750,17 @@ func TestCreateNodePoolWorkflowExecute(t *testing.T) {
 			environment.RegisterActivityWithOptions(
 				func(
 					ctx context.Context,
-					input awsworkflow.CalculateNodePoolVersionActivityInput,
-				) (*awsworkflow.CalculateNodePoolVersionActivityOutput, error) {
+					input CalculateNodePoolVersionActivityInput,
+				) (*CalculateNodePoolVersionActivityOutput, error) {
 					if testCase.expectedError != nil &&
 						strings.HasPrefix(testCase.expectedError.Error(), "node pool calculation error") {
 						return nil, testCase.expectedError
 					}
 
-					return &awsworkflow.CalculateNodePoolVersionActivityOutput{}, nil
+					return &CalculateNodePoolVersionActivityOutput{}, nil
 				},
 				activity.RegisterOptions{
-					Name: awsworkflow.CalculateNodePoolVersionActivityName,
+					Name: CalculateNodePoolVersionActivityName,
 				},
 			)
