--- internal/cluster/infrastructure/aws/awsworkflow/BUILD
@@ -1,21 +1,0 @@
-subinclude("///pleasings2//go:compat")
-
-go_library(
-    name = "awsworkflow",
-    srcs = glob(
-        ["*.go"],
-        exclude = ["*_test.go"],
-    ),
-    visibility = ["PUBLIC"],
-    deps = [
-        "//internal/cluster",
-        ],
-)
-
-go_test(
-    name = "test",
-    srcs = glob(["*_test.go"]),
-    deps = [
-        ":awsworkflow",
-    ],
-)
